requirements.txt
requests==2.32.3
tenacity==8.5.0
SQLAlchemy==2.0.32
pyodbc==5.1.0
pydantic==2.8.2
pandas==2.2.2
python-dateutil==2.9.0.post0
apscheduler==3.10.4
python-dotenv==1.0.1
stomp.py==8.1.0
retry==0.9.2
tqdm==4.66.4
fastapi==0.112.2
uvicorn==0.30.6
jinja2==3.1.4
